+++ sdk/search/azure-search-documents/azure/search/documents/_service/_search_service_client.py
@@ -7,6 +7,7 @@ from typing import TYPE_CHECKING
 
 from azure.core.tracing.decorator import distributed_trace
 
+from .search_service_client_base import SearchServiceClientBase
 from ._generated import SearchServiceClient as _SearchServiceClient
 from .._headers_mixin import HeadersMixin
 from .._version import SDK_MONIKER
@@ -22,7 +23,7 @@ if TYPE_CHECKING:
     from azure.core.credentials import AzureKeyCredential
 
 
-class SearchServiceClient(HeadersMixin):  # pylint: disable=too-many-public-methods
+class SearchServiceClient(SearchServiceClientBase):  # pylint: disable=too-many-public-methods
     """A client to interact with an existing Azure search service.
 
     :param endpoint: The URL endpoint of an Azure search service
@@ -44,21 +45,7 @@ class SearchServiceClient(HeadersMixin):  # pylint: disable=too-many-public-meth
 
     def __init__(self, endpoint, credential, **kwargs):
         # type: (str, AzureKeyCredential, **Any) -> None
-
-        try:
-            if endpoint.lower().startswith('http') and not endpoint.lower().startswith('https'):
-                raise ValueError("Endpoint should be secure. Use https.")
-            if not endpoint.lower().startswith('http'):
-                endpoint = "https://" + endpoint
-        except AttributeError:
-            raise ValueError("Endpoint must be a string")
-
-        self._endpoint = endpoint  # type: str
-        self._credential = credential  # type: AzureKeyCredential
-        self._client = _SearchServiceClient(
-            endpoint=endpoint, sdk_moniker=SDK_MONIKER, **kwargs
-        )  # type: _SearchServiceClient
-
+        super(SearchServiceClient, self).__init__(endpoint, credential, **kwargs)
         self._indexes_client = SearchIndexesClient(endpoint, credential, **kwargs)
 
         self._synonym_maps_client = SearchSynonymMapsClient(
@@ -72,10 +59,6 @@ class SearchServiceClient(HeadersMixin):  # pylint: disable=too-many-public-meth
         )
 
         self._indexers_client = SearchIndexersClient(endpoint, credential, **kwargs)
-
-    def __repr__(self):
-        # type: () -> str
-        return "<SearchServiceClient [endpoint={}]>".format(repr(self._endpoint))[:1024]
 
     def __enter__(self):
         # type: () -> SearchServiceClient
